(PC-10774) re-organize firestore code
