+++ src/libs/firestore/client.ts
@@ -1,0 +1,4 @@
+import firestore from '@react-native-firebase/firestore'
+
+const firestoreRemoteStore = firestore()
+export default firestoreRemoteStore
