+++ src/libs/firestore/types.ts
@@ -1,0 +1,7 @@
+export enum RemoteStoreCollections {
+  MAINTENANCE = 'maintenance',
+}
+
+export enum RemoteStoreDocuments {
+  MAINTENANCE_IS_ON = 'maintenanceIsOn',
+}
